use std::sync::Condvar to wait for keys being pressed instead of busy looping which is using 100% CPU.
std::sync::Condvar

Condition variables represent the ability to block a thread such that it consumes no CPU time while waiting for an event to occur.
Condition variables are typically associated with a boolean predicate (a condition) and a mutex.
The predicate is always verified inside of the mutex before determining that a thread must block.

Functions in this module will block the current thread of execution.
Note that any attempt to use multiple mutexes on the same condition variable may result in a runtime panic.

https://doc.rust-lang.org/std/sync/struct.Condvar.html